+++ src/Cdn/exports/New-AzFrontDoorCdnRule.ps1
@@ -49,7 +49,7 @@ ACTION <IDeliveryRuleAction1[]>: A list of actions that are executed when all th
 CONDITION <IDeliveryRuleCondition[]>: A list of conditions that must be matched for the actions to be executed
   Name <MatchVariable>: The name of the condition for the delivery rule.
 .Link
-https://docs.microsoft.com/powershell/module/az.cdn/new-azfrontdoorcdnrule
+https://learn.microsoft.com/powershell/module/az.cdn/new-azfrontdoorcdnrule
 #>
 function New-AzFrontDoorCdnRule {
 [OutputType([Microsoft.Azure.PowerShell.Cmdlets.Cdn.Models.Api20210601.IRule])]